Authentication and Authorization commit 1
Added User Schema  model to mongoose  for creating and managing the data of the users. Created auth controller for authentication of the users and implemented the logic according to User model. User bcrpyt to hash the password. Added register middleware to manage the post requests coming to register route. Added pictures that I will use in the app.