Friend component, PostWidget and PostsWidget widgets created
The Friend components helps me to display the users that posted the posts or our friends in the widget I will create that displays our friends. PostWidget displays the post with the parameters it receive from PostsWidget. PostsWidget is displaying the feedPosts or userPosts depending on the page it is rendered.